Upgrade version of commander dependency
A new version of the [commander](https://www.npmjs.com/package/commander) package is available. However, there are no functional changes in the release; it is simply a repackaging with an older version of npm to get around a [bug](https://github.com/npm/npm/issues/19968) that had been introduced in npm 5.7.1. This lack of functional changes can be verified by comparing the differences in the codebase between [v2.15.0 and v2.15.1](https://github.com/tj/commander.js/compare/v2.15.0...v2.15.1). Though nothing has changed, it is important to keep the dependency's version number up to date.